# Collect memory optimization