--- a/test/test_files/agg/hash_ldbc.test
+++ b/test/test_files/agg/hash_ldbc.test
@@ -48,3 +48,9 @@ Internet Explorer|[Chrome,Firefox,Internet Explorer,Opera,Safari]
 Chrome|[Chrome,Firefox,Internet Explorer,Opera,Safari]
 Safari|[Chrome,Firefox,Internet Explorer,Opera,Safari]
 Opera|[Chrome,Firefox,Internet Explorer,Opera,Safari]
+
+-CASE Debug
+-STATEMENT MATCH (a:Comment)-[e:replyOf_Comment]->(b:Comment)
+            WITH b, collect(a) as aList
+            RETURN b, aList
+---- ok